backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
httpx==0.27.0
python-dateutil==2.9.0
ollama==0.3.0
google-auth==2.29.0
google-cloud-storage==2.16.0
